Surface rendering: merge *Uniforms.ts into *Shader.ts
Six paired files (TerrainScreen, TerrainComposite, TerrainTile,
WaterHeight, WaterFilter, WindField) had their defineUniformStruct
call sit in a separate ~30-line Uniforms file beside the Shader file
that already embedded the same struct as a WGSL preamble string. Two
sources of truth, no benefit.

Folded each Uniforms call into its Shader file, deleted the standalone
files, and updated SurfaceRenderer/TerrainTileCache/WaterTuning to
import from the Shader module. BoatAir and WetnessState already had
their uniforms inline.
