+++ src/game/surface-rendering/WaterHeightShader.ts
@@ -11,6 +11,12 @@
  */
 
 import {
+  defineUniformStruct,
+  f32,
+  mat3x3,
+  u32,
+} from "../../core/graphics/UniformStruct";
+import {
   ComputeShader,
   type ComputeShaderConfig,
 } from "../../core/graphics/webgpu/ComputeShader";
@@ -27,6 +33,30 @@ import {
 } from "../world/water/WaterConstants";
 
 const WORKGROUP_SIZE = [8, 8] as const;
+
+/**
+ * Uniforms for the water height compute pass.
+ */
+export const WaterHeightUniforms = defineUniformStruct("Params", {
+  // Clip-space → world transform for the output texture.
+  // Same layout as TerrainScreenUniforms.texClipToWorld.
+  texClipToWorld: mat3x3,
+
+  // Dimensions of the output texture in texels.
+  textureWidth: f32,
+  textureHeight: f32,
+
+  // Time and water params
+  time: f32,
+  tideHeight: f32,
+
+  // Wave configuration (from level data)
+  numWaves: u32,
+
+  // Multiplier on Gerstner wave amplitude. 1.0 = no change. Driven from
+  // `WeatherState.waveAmplitudeScale`.
+  waveAmplitudeScale: f32,
+});
 
 /**
  * Params module with uniforms and bindings for water height computation.
