+++ src/game/surface-rendering/WindFieldShader.ts
@@ -9,6 +9,12 @@
  */
 
 import {
+  defineUniformStruct,
+  f32,
+  mat3x3,
+  u32,
+} from "../../core/graphics/UniformStruct";
+import {
   ComputeShader,
   type ComputeShaderConfig,
 } from "../../core/graphics/webgpu/ComputeShader";
@@ -26,6 +32,28 @@ import {
 } from "../world/wind/WindConstants";
 
 const WORKGROUP_SIZE = [8, 8] as const;
+
+export const WindFieldUniforms = defineUniformStruct("Params", {
+  texClipToWorld: mat3x3,
+
+  textureWidth: u32,
+  textureHeight: u32,
+
+  time: f32,
+  baseWindX: f32,
+  baseWindY: f32,
+
+  numActiveWindSources: u32,
+
+  weights0: f32,
+  weights1: f32,
+  weights2: f32,
+  weights3: f32,
+  weights4: f32,
+  weights5: f32,
+  weights6: f32,
+  weights7: f32,
+});
 
 const windFieldParamsModule: ShaderModule = {
   preamble: /*wgsl*/ `
